Implement more menu features
Menus for linking and unlinking* records.
Remove menus for vectors.
New deleteElement() interface for vectors (specify position to delete
instead of element).
Add option to define a function for additional preparations when adding
a new element.

*linking and unlinking - refers to adding and removing elements in the
vectors containing keys to a map in the school (e.g. a teacher has a
vector with team IDs)